Break apart batch uploader tests to improve test reliability
