+++ test/src/_test.index.ts
@@ -17,6 +17,7 @@ import './tests-identities-attributes';
 import './tests-native-sdk';
 import './tests-consent';
 import './tests-serverModel';
+import './tests-batchUploader_2';
 import './tests-mockBatchCreator';
 import './tests-mParticleUser';
 import './tests-self-hosting-specific';
@@ -24,6 +25,7 @@ import './tests-runtimeToBatchEventsDTO';
 import './tests-apiClient';
 import './tests-mparticle-instance-manager';
 import './tests-queue-public-methods';
+import './tests-batchUploader_3';
 import './tests-validators';
 import './tests-utils';
 import './tests-session-manager';
@@ -35,3 +37,5 @@ import './tests-feature-flags';
 import './tests-user';
 import './tests-legacy-alias-requests';
 import './tests-aliasRequestApiClient';
+import './tests-batchUploader_4';
+
