+++ test/src/tests-batchUploader.ts
@@ -1116,952 +1116,5 @@ describe('batch uploader', () => {
         .catch((e) => {
         })
         });
-    describe('Upload Workflow', () => {
-        beforeEach(() => {
-        });
-
-        afterEach(() => {
-            fetchMock.restore();
-        });
-
-        it('should organize events in the order they are processed and maintain that order when uploading', (done) => {
-            // Batches should be uploaded in the order they were created to prevent
-            // any potential corruption.
-            fetchMock.post(urls.events, 200);
-            fetchMock.config.overwriteRoutes = true;
-
-            window.mParticle.config.flags = {
-                ...enableBatchingConfigFlags,
-            };
-
-            window.mParticle._resetForTests(MPConfig);
-            window.mParticle.init(apiKey, window.mParticle.config);
-            waitForCondition(hasIdentifyReturned)
-            .then(() => {
-            window.mParticle.logEvent('Test Event 0');
-
-            // Manually initiate the upload process - turn event into batches and upload the batch
-            window.mParticle.upload();
-
-            expect(
-                fetchMock.called(),
-                'FetchMock should have been called'
-            ).to.equal(true);
-
-            const batch1 = JSON.parse(fetchMock.lastCall()[1].body as string);
-
-            // Batch 1 should contain only session start, AST and a single event
-            // in this exact order
-            expect(batch1.events.length).to.equal(3);
-            expect(batch1.events[0].event_type).to.equal('session_start');
-            expect(batch1.events[1].event_type).to.equal(
-                'application_state_transition'
-            );
-            expect(batch1.events[2].data.event_name).to.equal('Test Event 0');
-
-            // Log a second batch of events
-            window.mParticle.logEvent('Test Event 1');
-            window.mParticle.logEvent('Test Event 2');
-            window.mParticle.logEvent('Test Event 3');
-
-            // Manually initiate the upload process - turn event into batches and upload the batch 
-            window.mParticle.upload();
-
-            const batch2 = JSON.parse(fetchMock.lastCall()[1].body as string);
-
-            // Batch 2 should contain three custom events
-            expect(batch2.events.length).to.equal(3);
-            expect(batch2.events[0].data.event_name).to.equal('Test Event 1');
-            expect(batch2.events[1].data.event_name).to.equal('Test Event 2');
-            expect(batch2.events[2].data.event_name).to.equal('Test Event 3');
-
-            // Log a third batch of events
-            window.mParticle.logEvent('Test Event 4');
-            window.mParticle.logEvent('Test Event 5');
-
-            // Manually initiate the upload process - turn event into batches and upload the batch 
-            window.mParticle.upload();
-
-            const batch3 = JSON.parse(fetchMock.lastCall()[1].body as string);
-
-            // Batch 3 should contain two custom events
-            expect(batch3.events.length).to.equal(2);
-            expect(batch3.events[0].data.event_name).to.equal('Test Event 4');
-            expect(batch3.events[1].data.event_name).to.equal('Test Event 5');
-
-            done();
-
-        })
-        .catch(done);
-        });
-
-        // TODO: Investigate workflow with unshift vs push
-        // https://go.mparticle.com/work/SQDSDKS-5165
-        it.skip('should keep batches in sequence for future retries when an HTTP 500 error occurs', (done) => {
-            // If batches cannot upload, they should be added back to the Batch Queue
-            // in the order they were created so they can be retransmitted.
-
-            fetchMock.post(urls.events, 500);
-
-            window.mParticle.config.flags = {
-                ...enableBatchingConfigFlags,
-            };
-
-            window.mParticle._resetForTests(MPConfig);
-            window.mParticle.init(apiKey, window.mParticle.config);
-            // Generates Batch 1 with Session Start + AST
-
-            // Adds a custom event to Batch 1
-            window.mParticle.logEvent('Test Event 0');
-
-            // Manually initiate the upload process - turn event into batches and upload the batch 
-            window.mParticle.upload();
-
-            expect(
-                fetchMock.called(),
-                'FetchMock should have been called'
-            ).to.equal(true);
-
-            const batch1 = JSON.parse(fetchMock.calls()[0][1].body as string);
-
-            // Batch 1 should contain only session start, AST and a single event
-            // in this exact order
-            expect(batch1.events.length).to.equal(3);
-            expect(batch1.events[0].event_type).to.equal('session_start');
-            expect(batch1.events[1].event_type).to.equal(
-                'application_state_transition'
-            );
-            expect(batch1.events[2].data.event_name).to.equal('Test Event 0');
-
-            // Batch 2
-            window.mParticle.logEvent('Test Event 1');
-            window.mParticle.logEvent('Test Event 2');
-            window.mParticle.logEvent('Test Event 3');
-
-            // Manually initiate the upload process - turn event into batches and upload the batch
-            window.mParticle.upload();
-
-            // Batch 3
-            window.mParticle.logEvent('Test Event 4');
-            window.mParticle.logEvent('Test Event 5');
-            window.mParticle.logEvent('Test Event 6');
-
-            // Manually initiate the upload process - turn event into batches and upload the batch
-            window.mParticle.upload();
-
-            // Reset clock so that the setTimeout return immediately
-            clock.restore();
-
-            setTimeout(() => {
-                const batchQueue =
-                    window.mParticle.getInstance()._APIClient.uploader
-                        .batchesQueuedForProcessing;
-
-                expect(batchQueue.length).to.equal(3);
-
-                expect(batchQueue[0].events[0].event_type).to.equal(
-                    'session_start'
-                );
-                expect(batchQueue[0].events[1].event_type).to.equal(
-                    'application_state_transition'
-                );
-                expect(batchQueue[0].events[2].data.event_name).to.equal('Test Event 0');
-
-                expect(batchQueue[1].events[0].data.event_name).to.equal(
-                    'Test Event 1'
-                );
-                expect(batchQueue[1].events[1].data.event_name).to.equal(
-                    'Test Event 2'
-                );
-                expect(batchQueue[1].events[2].data.event_name).to.equal(
-                    'Test Event 3'
-                );
-
-                expect(batchQueue[2].events[0].data.event_name).to.equal(
-                    'Test Event 4'
-                );
-                expect(batchQueue[2].events[1].data.event_name).to.equal(
-                    'Test Event 5'
-                );
-                expect(batchQueue[2].events[2].data.event_name).to.equal(
-                    'Test Event 6'
-                );
-
-                done();
-            }, 0);
-        });
-
-        // TODO: Investigate workflow with unshift vs push
-        // https://go.mparticle.com/work/SQDSDKS-5165
-        it.skip('should keep and retry batches in sequence if the transmission fails midway', (done) => {
-            // First request is successful, subsequent requests fail
-            fetchMock.post(urls.events, 200, {
-                overwriteRoutes: false,
-                repeat: 1,
-            });
-
-            fetchMock.post(urls.events, 429, {
-                overwriteRoutes: false,
-            });
-
-            window.mParticle.config.flags = {
-                ...enableBatchingConfigFlags,
-            };
-
-            window.mParticle._resetForTests(MPConfig);
-            window.mParticle.init(apiKey, window.mParticle.config);
-            // Generates Batch 1 with Session Start + AST
-
-            // Adds a custom event to Batch 1
-            window.mParticle.logEvent('Test Event 0');
-
-            // Manually initiate the upload process - turn event into batches and upload the batch 
-            window.mParticle.upload();
-
-            // Batch 2
-            window.mParticle.logEvent('Test Event 1');
-            window.mParticle.logEvent('Test Event 2');
-            window.mParticle.logEvent('Test Event 3');
-
-            // Manually initiate the upload process - turn event into batches and upload the batch 
-            window.mParticle.upload();
-
-            // Batch 3
-            window.mParticle.logEvent('Test Event 4');
-            window.mParticle.logEvent('Test Event 5');
-            window.mParticle.logEvent('Test Event 6');
-
-            // Manually initiate the upload process - turn event into batches and upload the batch 
-            window.mParticle.upload();
-
-            // Reset timer so the setTimeout can trigger
-            clock.restore();
-
-            setTimeout(() => {
-                // Batch upload should be triggered 3 times, but only
-                // 2 should be waiting for retry
-                expect(fetchMock.calls().length).to.equal(3);
-
-                const batchQueue =
-                    window.mParticle.getInstance()._APIClient.uploader
-                        .batchesQueuedForProcessing;
-
-                expect(batchQueue.length).to.equal(2);
-
-                expect(batchQueue[0].events[0].data.event_name).to.equal(
-                    'Test Event 1'
-                );
-                expect(batchQueue[0].events[1].data.event_name).to.equal(
-                    'Test Event 2'
-                );
-                expect(batchQueue[0].events[2].data.event_name).to.equal(
-                    'Test Event 3'
-                );
-
-                expect(batchQueue[1].events[0].data.event_name).to.equal(
-                    'Test Event 4'
-                );
-                expect(batchQueue[1].events[1].data.event_name).to.equal(
-                    'Test Event 5'
-                );
-                expect(batchQueue[1].events[2].data.event_name).to.equal(
-                    'Test Event 6'
-                );
-
-                done();
-            }, 0);
-        });
-    });
-
-    describe('batching via window.fetch', () => {
-        beforeEach(() => {
-            fetchMock.post(urls.events, 200);
-            fetchMock.config.overwriteRoutes = true;
-            window.mParticle.config.flags.eventBatchingIntervalMillis = 1000;
-        });
-
-        afterEach(() => {
-            fetchMock.restore();
-            window.mParticle.config.flags.eventBatchingIntervalMillis = 0;
-        });
-
-        it('should use custom v3 endpoint', function(done) {
-            window.mParticle._resetForTests(MPConfig);
-            fetchMock.resetHistory();
-            window.mParticle.init(apiKey, window.mParticle.config);
-            waitForCondition(hasIdentifyReturned)
-            .then(() => {
-            window.mParticle.logEvent('Test Event');
-            
-            window.mParticle.upload();
-
-            const lastCall = fetchMock.lastCall();
-            const endpoint = lastCall[0];
-            const batch = JSON.parse(fetchMock.lastCall()[1].body as string);
-
-            endpoint.should.equal(urls.events);
-            batch.events[0].event_type.should.equal('session_start');
-            batch.events[1].event_type.should.equal('application_state_transition');
-            batch.events[2].event_type.should.equal('custom_event');
-            batch.events[2].data.event_name.should.equal('Test Event');
-    
-            done();
-            })
-        });
-
-        it('should have latitude/longitude for location when batching', function(done) {
-            window.mParticle._resetForTests(MPConfig);
-            window.mParticle.init(apiKey, window.mParticle.config);
-            waitForCondition(hasIdentifyReturned)
-            .then(() => {
-            window.mParticle.setPosition(100, 100);
-            window.mParticle.logEvent('Test Event');
-            window.mParticle.upload();
-            const lastCall = fetchMock.lastCall();
-            const endpoint = lastCall[0];
-            const batch = JSON.parse(fetchMock.lastCall()[1].body as string);
-            endpoint.should.equal(urls.events);
-            batch.events[2].data.location.should.have.property('latitude', 100)
-            batch.events[2].data.location.should.have.property('longitude', 100)
-    
-            done();
-            })
-        });
-
-        it('should force uploads when using public `upload`', function(done) {
-            window.mParticle._resetForTests(MPConfig);
-            window.mParticle.init(apiKey, window.mParticle.config);
-            waitForCondition(hasIdentifyReturned)
-            .then(() => {
-            window.mParticle.logEvent('Test Event');
-            // Identity call
-            // Session start, AST, and `Test Event` are queued.
-            fetchMock.calls().length.should.equal(1);
-            (fetchMock.lastCall()[0].endsWith('identify')).should.equal(true)
-            
-            // force upload, triggering window.fetch
-            window.mParticle.upload();
-
-            const lastCall = fetchMock.lastCall();
-            const endpoint = lastCall[0];
-            const batch = JSON.parse(fetchMock.lastCall()[1].body as string);
-
-            endpoint.should.equal(urls.events);
-            batch.events[0].event_type.should.equal('session_start');
-            batch.events[1].event_type.should.equal('application_state_transition');
-            batch.events[2].event_type.should.equal('custom_event');
-            batch.events[2].data.event_name.should.equal('Test Event');
-
-            done();
-            })
-        });
-
-        it('should force uploads when a commerce event is called', function(done) {
-            window.mParticle._resetForTests(MPConfig);
-            window.mParticle.init(apiKey, window.mParticle.config);
-
-            waitForCondition(hasIdentifyReturned)
-            .then(() => {
-            window.mParticle.logEvent('Test Event');
-
-            var product1 = window.mParticle.eCommerce.createProduct('iphone', 'iphoneSKU', 999);
-            window.mParticle.eCommerce.logProductAction(SDKProductActionType.AddToCart, product1);
-
-            const lastCall = fetchMock.lastCall();
-            const endpoint = lastCall[0];
-            const batch = JSON.parse(fetchMock.lastCall()[1].body as string);
-
-            endpoint.should.equal(urls.events);
-            batch.events[0].event_type.should.equal('session_start');
-            batch.events[1].event_type.should.equal('application_state_transition');
-            batch.events[2].event_type.should.equal('custom_event');
-            batch.events[2].data.event_name.should.equal('Test Event');
-            batch.events[3].event_type.should.equal('commerce_event');
-            batch.events[3].data.product_action.action.should.equal('add_to_cart');
-
-            done();
-            });
-        });
-
-        it('should return pending uploads if a 500 is returned', function(done) {
-            window.mParticle._resetForTests(MPConfig);
-
-            fetchMock.post(urls.events, 500);
-            
-            window.mParticle.init(apiKey, window.mParticle.config);
-            waitForCondition(hasIdentifyReturned)
-            .then(() => {
-            window.mParticle.logEvent('Test Event');
-
-            let pendingEvents = window.mParticle.getInstance()._APIClient.uploader.eventsQueuedForProcessing;
-
-            pendingEvents.length.should.equal(3)
-            pendingEvents[0].EventName.should.equal(1);
-            pendingEvents[1].EventName.should.equal(10);
-            pendingEvents[2].EventName.should.equal('Test Event');
-
-            fetchMock.post(urls.events, 200);
-            
-            // First fetch call is an identify call
-            (fetchMock.lastCall()[0].endsWith('identify')).should.equal(true)
-            window.mParticle.upload();
-
-            let nowPendingEvents = window.mParticle.getInstance()._APIClient.uploader.eventsQueuedForProcessing;
-            nowPendingEvents.length.should.equal(0);
-
-            const batch = JSON.parse(fetchMock.lastCall()[1].body as string);
-            batch.events[0].event_type.should.equal('session_start');
-            batch.events[1].event_type.should.equal('application_state_transition');
-            batch.events[2].event_type.should.equal('custom_event');
-            batch.events[2].data.event_name.should.equal('Test Event');
-
-            done();
-            })
-        });
-
-        it('should send source_message_id with events to v3 endpoint', function(done) {
-            window.mParticle._resetForTests(MPConfig);
-            window.mParticle.init(apiKey, window.mParticle.config);
-
-            waitForCondition(hasIdentifyReturned)
-            .then(() => {
-            window.mParticle.logEvent('Test Event');
-
-            window.mParticle.upload();
-
-            const lastCall = fetchMock.lastCall();
-            const endpoint = lastCall[0];
-            const batch = JSON.parse(fetchMock.lastCall()[1].body as string);
-
-            endpoint.should.equal(urls.events);
-            batch.events[0].data.should.have.property('source_message_id')
-
-            done();
-            })
-        });
-
-        it('should send user-defined SourceMessageId events to v3 endpoint', function(done) {
-            window.mParticle._resetForTests(MPConfig);
-            window.mParticle.init(apiKey, window.mParticle.config);
-            waitForCondition(hasIdentifyReturned)
-            .then(() => {
-            window.mParticle.logBaseEvent({
-                messageType: 4,
-                name: 'Test Event',
-                data: {key: 'value'},
-                eventType: 3,
-                customFlags: {flagKey: 'flagValue'},
-                sourceMessageId: 'abcdefg'
-            });
-
-            window.mParticle.upload();
-
-            const lastCall = fetchMock.lastCall();
-            const endpoint = lastCall[0];
-            const batch = JSON.parse(fetchMock.lastCall()[1].body as string);
-
-            endpoint.should.equal(urls.events);
-            // event batch includes session start, ast, then last event is Test Event
-            batch.events[batch.events.length-1].data.should.have.property('source_message_id', 'abcdefg')
-
-            done();
-            })
-        });
-
-        it('should call the identity callback after a session ends if user is returning to the page after a long period of time', async function() {
-            // Background of bug that this test fixes:
-            // User navigates away from page and returns after some time
-            // and the session should end.  There is a UAC firing inside of
-            // config.identityCallback, which would send to our servers with
-            // the previous session ID because the identity call back fired
-            // before the session logic that determines if a new session should
-            // start.
-
-            window.mParticle._resetForTests(MPConfig);
-            
-            window.mParticle.config.identityCallback = function(result) {
-                let currentUser = result.getUser()
-                if (currentUser) {
-                    // TODO: Investigate if we should update definitely typed typings for
-                    // setUserAttribute which only allows strings right now
-                    // more context at https://go.mparticle.com/work/SQDSDKS-4576
-                    currentUser.setUserAttribute("number", `${Math.floor((Math.random() * 1000) + 1)}`)
-                }
-            }
-            
-            var endSessionFunction = window.mParticle.getInstance()._SessionManager.endSession;
-            
-            window.mParticle.init(apiKey, window.mParticle.config);
-            waitForCondition(() => {
-            return (
-                window.mParticle.getInstance()._Store.identityCallInFlight === false
-            );
-            })
-            .then(() => {
-            // Mock end session so that the SDK doesn't actually send it. We do this
-            // to mimic a return to page behavior, below:
-            window.mParticle.getInstance()._SessionManager.endSession = function() {}
-            clock = sinon.useFakeTimers({now: new Date().getTime()});
-
-            // Force 35 minutes to pass, so that when we return to the page, when
-            // the SDK initializes it will know to end the session.
-            clock.tick(35*60000);
-
-            // Undo mock of end session so that when we initializes, it will end
-            // the session for real.
-            window.mParticle.getInstance()._SessionManager.endSession = endSessionFunction;
-
-            // Initialize imitates returning to the page
-            window.mParticle.init(apiKey, window.mParticle.config);
-            waitForCondition(() => {
-            return (
-                window.mParticle.getInstance()._Store.identityCallInFlight === false
-            );
-            })
-            .then(async () => {
-            // Manually initiate the upload process - turn event into batches and upload the batch 
-            await window.mParticle.getInstance()._APIClient.uploader.prepareAndUpload();
-
-            const batch1 = JSON.parse(fetchMock.calls()[0][1].body as string);
-            const batch2 = JSON.parse(fetchMock.calls()[1][1].body as string);
-            const batch3 = JSON.parse(fetchMock.calls()[2][1].body as string);
-
-            // UAC, session start, AST
-            expect(
-                batch1.events.length,
-                'Batch 1: UAC Event, Session Start, AST'
-            ).to.equal(3);
-
-            // session end
-            expect(batch2.events.length, 'Batch 2: Session End').to.equal(
-                1
-            );
-
-            // UAC, session start, AST
-            expect(
-                batch3.events.length,
-                'Batch 3: UAC, Session Start, AST'
-            ).to.equal(3);
-
-            const batch1UAC = Utils.findEventFromBatch(
-                batch1,
-                'user_attribute_change'
-            );
-            const batch1SessionStart = Utils.findEventFromBatch(
-                batch1,
-                'session_start'
-            );
-            const batch1AST = Utils.findEventFromBatch(
-                batch1,
-                'application_state_transition'
-            );
-
-            batch1UAC.should.be.ok();
-            batch1SessionStart.should.be.ok();
-            batch1AST.should.be.ok();
-
-            const batch2SessionEnd = Utils.findEventFromBatch(
-                batch2,
-                'session_end'
-            );
-            batch2SessionEnd.should.be.ok();
-
-            const batch3UAC = Utils.findEventFromBatch(
-                batch3,
-                'user_attribute_change'
-            );
-            const batch3SessionStart = Utils.findEventFromBatch(
-                batch3,
-                'session_start'
-            );
-            const batch3AST = Utils.findEventFromBatch(
-                batch3,
-                'application_state_transition'
-            );
-
-            batch3UAC.should.be.ok();
-            batch3SessionStart.should.be.ok();
-            batch3AST.should.be.ok();
-
-
-            (typeof batch1.source_request_id).should.equal('string');
-            (typeof batch2.source_request_id).should.equal('string');
-            (typeof batch3.source_request_id).should.equal('string');
-
-            batch1.source_request_id.should.not.equal(
-                batch2.source_request_id
-            );
-            batch1.source_request_id.should.not.equal(
-                batch3.source_request_id
-            );
-            batch2.source_request_id.should.not.equal(
-                batch3.source_request_id
-            );
-
-            batch1UAC.data.session_uuid.should.equal(
-                batch1AST.data.session_uuid
-            );
-            batch1UAC.data.session_uuid.should.equal(
-                batch1SessionStart.data.session_uuid
-            );
-            batch1UAC.data.session_uuid.should.not.equal(
-                batch3UAC.data.session_uuid
-            );
-            batch1UAC.data.session_uuid.should.not.equal(
-                batch3SessionStart.data.session_uuid
-            );
-            batch1UAC.data.session_uuid.should.not.equal(
-                batch3AST.data.session_uuid
-            );
-
-            batch1UAC.data.session_start_unixtime_ms.should.equal(
-                batch1AST.data.session_start_unixtime_ms
-            );
-            batch1UAC.data.session_start_unixtime_ms.should.equal(
-                batch1SessionStart.data.session_start_unixtime_ms
-            );
-            batch1UAC.data.session_start_unixtime_ms.should.not.equal(
-                batch3UAC.data.session_start_unixtime_ms
-            );
-            batch1UAC.data.session_start_unixtime_ms.should.not.equal(
-                batch3SessionStart.data.session_start_unixtime_ms
-            );
-            batch1UAC.data.session_start_unixtime_ms.should.not.equal(
-                batch3AST.data.session_start_unixtime_ms
-            );
-
-            batch1SessionStart.data.session_uuid.should.equal(
-                batch1AST.data.session_uuid
-            );
-            batch1SessionStart.data.session_uuid.should.equal(
-                batch2SessionEnd.data.session_uuid
-            );
-            batch1AST.data.session_uuid.should.equal(
-                batch2SessionEnd.data.session_uuid
-            );
-
-            batch1SessionStart.data.session_start_unixtime_ms.should.equal(
-                batch1AST.data.session_start_unixtime_ms
-            );
-            batch1SessionStart.data.session_start_unixtime_ms.should.equal(
-                batch2SessionEnd.data.session_start_unixtime_ms
-            );
-            batch1AST.data.session_start_unixtime_ms.should.equal(
-                batch2SessionEnd.data.session_start_unixtime_ms
-            );
-
-            batch3AST.data.session_uuid.should.equal(
-                batch3UAC.data.session_uuid
-            );
-            batch3SessionStart.data.session_uuid.should.equal(
-                batch3UAC.data.session_uuid
-            );
-            batch3SessionStart.data.session_uuid.should.equal(
-                batch3AST.data.session_uuid
-            );
-
-            batch3AST.data.session_start_unixtime_ms.should.equal(
-                batch3UAC.data.session_start_unixtime_ms
-            );
-            batch3SessionStart.data.session_start_unixtime_ms.should.equal(
-                batch3UAC.data.session_start_unixtime_ms
-            );
-            batch3SessionStart.data.session_start_unixtime_ms.should.equal(
-                batch3AST.data.session_start_unixtime_ms
-            );
-
-            clock.restore();
-            })
-            })
-        });
-    });
-
-    describe('batching via XHR for older browsers without window.fetch', () => {
-        var fetch = window.fetch;
-
-        beforeEach(() => {
-            delete window.fetch
-            window.mParticle.config.flags = {
-                eventBatchingIntervalMillis: 1000,
-            }
-            mockServer = sinon.createFakeServer();
-            mockServer.respondImmediately = true;
-
-            mockServer.respondWith(urls.events, [
-                200,
-                {},
-                JSON.stringify({ mpid: testMPID, Store: {}})
-            ]);
-            mockServer.respondWith(urls.identify, [
-                200,
-                {},
-                JSON.stringify({ mpid: testMPID, is_logged_in: false }),
-            ]);
-        });
-    
-        afterEach(() => {
-            window.mParticle._resetForTests(MPConfig);
-            window.fetch = fetch;
-            sinon.restore();
-        });
-
-        it('should use custom v3 endpoint', function(done) {
-            window.mParticle._resetForTests(MPConfig);
-            mockServer.requests = [];
-            window.mParticle.init(apiKey, window.mParticle.config);
-            waitForCondition(hasIdentifyReturned)
-            .then(() => {
-            window.mParticle.logEvent('Test Event');
-            
-            mockServer.requests.length.should.equal(1);
-            window.mParticle.upload()
-            // 1st request is /Identity call, 2nd request is /Event call
-
-            const batch = JSON.parse(mockServer.secondRequest.requestBody);
-            
-            batch.events[0].event_type.should.equal('session_start');
-            batch.events[1].event_type.should.equal('application_state_transition');
-            batch.events[2].event_type.should.equal('custom_event');
-            batch.events[2].data.event_name.should.equal('Test Event');
-
-            done();
-            }).catch((e) => {
-            })
-        });
-
-        it('should force uploads when using public `upload`', function(done) {
-            window.mParticle._resetForTests(MPConfig);
-
-            window.mParticle.init(apiKey, window.mParticle.config);
-            waitForCondition(hasIdentifyReturned)
-            .then(() => {
-
-
-            window.mParticle.logEvent('Test Event');
-    
-            // The only request to the server should be the identify call
-            // Session start, AST, and Test Event are queued.
-            mockServer.requests.length.should.equal(1);
-            // Upload interval hit, now will send requests
-
-            window.mParticle.upload();
-            // 1st request is /Identity call, 2nd request is /Event call
-            const batch = JSON.parse(mockServer.secondRequest.requestBody);
-
-            batch.events[0].event_type.should.equal('session_start');
-            batch.events[1].event_type.should.equal('application_state_transition');
-            batch.events[2].event_type.should.equal('custom_event');
-            batch.events[2].data.event_name.should.equal('Test Event');
-
-            done();
-            })
-        });
-
-        it('should trigger an upload of batch when a commerce event is called', function(done) {
-            window.mParticle._resetForTests(MPConfig);
-
-            window.mParticle.init(apiKey, window.mParticle.config);
-            waitForCondition(hasIdentifyReturned)
-            .then(() => {
-            window.mParticle.logEvent('Test Event');
-            // The only request to the server should be the identify call
-            // Session start, AST, and Test Event are queued.
-            mockServer.requests.length.should.equal(1);
-
-            var product1 = window.mParticle.eCommerce.createProduct('iphone', 'iphoneSKU', 999);
-            window.mParticle.eCommerce.logProductAction(SDKProductActionType.AddToCart, product1);
-            // 1st request is /Identity call, 2nd request is /Event call
-            const batch = JSON.parse(mockServer.secondRequest.requestBody);
-
-            batch.events[0].event_type.should.equal('session_start');
-            batch.events[1].event_type.should.equal('application_state_transition');
-            batch.events[2].event_type.should.equal('custom_event');
-            batch.events[2].data.event_name.should.equal('Test Event');
-            batch.events[3].event_type.should.equal('commerce_event');
-            batch.events[3].data.product_action.action.should.equal('add_to_cart');
-
-            done();
-            })
-        });
-
-        it('should trigger an upload of batch when a UIC occurs', function(done) {
-            window.mParticle._resetForTests(MPConfig);
-            // include an identify request so that it creates a UIC
-            window.mParticle.config.identifyRequest = {
-                userIdentities: {
-                    customerid: 'foo-customer-id'
-                }
-            };
-
-            window.mParticle.init(apiKey, window.mParticle.config);
-            waitForCondition(hasIdentifyReturned)
-            .then(() => {
-            // Requests sent should be identify call, then UIC event
-            // Session start, AST, and Test Event are queued, and don't appear
-            // in the mockServer.requests
-            mockServer.requests.length.should.equal(2);
-
-            // 1st request is /Identity call, 2nd request is events call
-            const batch = JSON.parse(mockServer.secondRequest.requestBody);
-            
-            batch.events[2].event_type.should.equal('user_identity_change');
-
-
-            batch.events[0].event_type.should.equal('session_start');
-            batch.events[1].event_type.should.equal('application_state_transition');
-
-            done();
-            });
-        });
-
-        // Originally, we had the Batch uploader set to force an upload when a UAC event
-        // was triggered. This feature was removed and we are including this test to
-        // make sure the Web SDK does not regress. This test will be removed in a future
-        // Web SDK update
-        // TODO: https://go.mparticle.com/work/SQDSDKS-5891 
-        it('should NOT trigger an upload of batch when a UAC occurs', function(done) {
-            window.mParticle._resetForTests(MPConfig);
-
-            window.mParticle.init(apiKey, window.mParticle.config);
-            waitForCondition(hasIdentifyReturned)
-            .then(() => {
-
-            // Set a user attribute to trigger a UAC event
-            window.mParticle.Identity.getCurrentUser().setUserAttribute('age', 25);
-
-            // Requests sent should be identify call
-            // Since we no longer force an upload for UAC,
-            // This request will contain a /Identity Call
-            // a future request will contain session start, AST, and UAC
-            mockServer.requests.length.should.equal(1);
-
-            // Verifies that adding a UAC does not trigger an upload
-            expect(mockServer.secondRequest).to.equal(null);
-
-            // Manually force an upload
-            window.mParticle.upload();
-
-            // Second request has now been made
-            expect(mockServer.secondRequest).to.be.ok;
-            
-            const batch = JSON.parse(mockServer.secondRequest.requestBody);
-            
-            // Batch should now contain the 3 events we expect
-            mockServer.requests.length.should.equal(2);
-            batch.events[0].event_type.should.equal('session_start');
-            batch.events[1].event_type.should.equal('application_state_transition');
-            batch.events[2].event_type.should.equal('user_attribute_change');
-
-            done();
-        });
-        });
-
-        it('should return pending uploads if a 500 is returned', function(done) {
-            window.mParticle._resetForTests(MPConfig);
-
-            mockServer.respondWith(urls.events, [
-                500,
-                {},
-                JSON.stringify({ mpid: testMPID, Store: {}})
-            ]);
-
-            window.mParticle.init(apiKey, window.mParticle.config);
-
-            waitForCondition(hasIdentifyReturned)
-            .then(() => {
-            window.mParticle.logEvent('Test Event');
-
-            const pendingEvents = window.mParticle.getInstance()._APIClient.uploader.eventsQueuedForProcessing;
-
-            pendingEvents.length.should.equal(3)
-            pendingEvents[0].EventName.should.equal(1);
-            pendingEvents[1].EventName.should.equal(10);
-            pendingEvents[2].EventName.should.equal('Test Event');
-
-            window.mParticle.upload();
-
-            const nowPendingEvents = window.mParticle.getInstance()._APIClient.uploader.eventsQueuedForProcessing;
-            nowPendingEvents.length.should.equal(0);
-
-            const batch = JSON.parse(mockServer.secondRequest.requestBody);
-            batch.events[0].event_type.should.equal('session_start');
-            batch.events[1].event_type.should.equal('application_state_transition');
-            batch.events[2].event_type.should.equal('custom_event');
-            batch.events[2].data.event_name.should.equal('Test Event');
-            done();
-            });
-        });
-
-        it('should add a modified boolean of true to a batch that has been modified via a config.onCreateBatch call', function(done) {
-            window.mParticle._resetForTests(MPConfig);
-
-            window.mParticle.config.onCreateBatch = function (batch: Batch) {
-                return batch
-            };
-
-            window.mParticle.init(apiKey, window.mParticle.config);
-            waitForCondition(hasIdentifyReturned)
-            .then(() => {
-            window.mParticle.logEvent('Test Event');
-            
-            window.mParticle.upload()
-            
-            const batch = JSON.parse(mockServer.secondRequest.requestBody);
-            batch.modified.should.equal(true);
-            done();
-            });
-
-        });
-
-        it('should respect rules for the batch modification', function(done) {
-            window.mParticle._resetForTests(MPConfig);
-
-            window.mParticle.config.onCreateBatch = function (batch) {
-                batch.events.map(event => {
-                    if (event.event_type === "custom_event") {
-                        (event.data as CustomEventData).event_name = 'Modified!'
-                    }
-                });
-                return batch;
-            };
-
-            window.mParticle.init(apiKey, window.mParticle.config);
-
-            waitForCondition(hasIdentifyReturned)
-            .then(() => {
-            window.mParticle.logEvent('Test Event');
-            
-            window.mParticle.upload();
-
-            const batch = JSON.parse(mockServer.secondRequest.requestBody);
-            batch.events.length.should.equal(3);
-            batch.events[0].event_type.should.equal('session_start');
-            batch.events[1].event_type.should.equal('application_state_transition');
-            batch.events[2].event_type.should.equal('custom_event');
-            batch.events[2].data.event_name.should.equal('Modified!');
-            done();
-            });
-        });
-
-        it('should add a modified boolean of true to a batch that has been modified via a config.onCreateBatch call', function(done) {
-            window.mParticle._resetForTests(MPConfig);
-
-            window.mParticle.config.onCreateBatch = function (batch: Batch) {
-                return undefined;
-            };
-
-            window.mParticle.init(apiKey, window.mParticle.config);
-
-            waitForCondition(hasIdentifyReturned)
-            .then(() => {
-            window.mParticle.logEvent('Test Event');
-            
-            window.mParticle.upload();
-            
-            (mockServer.secondRequest === null).should.equal(true);
-            done();
-            })
-        });
-    });
     });
 });
